Renamed BaseOperationToolUI to BaseOperationChoiceToolUI
also fixed some imports
