+++ amulet_map_editor/programs/edit/api/ui/tool/__init__.py
@@ -1,3 +1,3 @@
-from .base_operation import BaseOperationToolUI
+from .base_operation_choice import BaseOperationChoiceToolUI
 from .base_tool_ui import BaseToolUI
 from .default_base_tool_ui import DefaultBaseToolUI
